Added separate files for dropdown and user authentication
