+++ frontend/src/components/DropdownButton.tsx
@@ -17,7 +17,17 @@ export const DropdownSelector = ({ title, options, onSelect, selectedValue }) =>
     <DropdownButton className="mt-5" title={title} onSelect={onSelect}>
     {options.map((option) => (
       <Dropdown.Item key={option} eventKey={option} active={selectedValue === option}>
-        {option}
+        {option.charAt(0).toUpperCase() + option.slice(1)}
+      </Dropdown.Item>
+    ))}
+  </DropdownButton>
+);
+
+export const DropdownSelectors = ({ title, options, onSelect }) => (
+  <DropdownButton title={title} onSelect={onSelect}>
+    {options.map((option, index) => (
+      <Dropdown.Item key={index} eventKey={option}>
+        {option.charAt(0).toUpperCase() + option.slice(1)}
       </Dropdown.Item>
     ))}
   </DropdownButton>
